+++ dev/user.clj
@@ -4,8 +4,7 @@
     [byte-streams :as bytes :refer [bytes=]]
     (blocks
       [core :as block]
-      [data :as data]
-      [util :as util])
+      [data :as data])
     (blocks.store
       [memory :refer [memory-store]]
       [file :refer [file-store]]
@@ -13,6 +12,9 @@
     [clojure.java.io :as io]
     [clojure.repl :refer :all]
     [clojure.string :as str]
+    [clojure.test.check :as check]
+    [clojure.test.check.generators :as gen]
+    [clojure.test.check.properties :as prop]
     [multihash.core :as multihash])
   (:import
     blocks.data.Block
@@ -22,6 +24,10 @@
 ; Conditional imports from clj-stacktrace and clojure.tools.namespace:
 (try (require '[clojure.stacktrace :refer [print-cause-trace]]) (catch Exception e nil))
 (try (require '[clojure.tools.namespace.repl :refer [refresh]]) (catch Exception e nil))
+
+
+(def test-blocks
+  (tests/generate-blocks! 10 1024))
 
 
 (def tbs