+++ test/blocks/block_test.clj
@@ -71,6 +71,6 @@
       (is (realized? literal))
       (is (not (realized? lazy))))
     (testing "deref logic"
-      (is (bytes= original @literal)
+      (is (some? @literal)
           "deref literal block should return content")
       (is (nil? @lazy) "deref lazy block returns nil"))))