+++ test/blocks/store/buffer_test.clj
@@ -4,7 +4,7 @@
     (blocks.store
       [memory :refer [memory-store]]
       [buffer :refer [buffer-store]]
-      [tests :refer [test-block-store]])
+      [tests :as tests])
     [clojure.test :refer :all]))
 
 
@@ -31,8 +31,4 @@
 
 
 (deftest ^:integration test-buffer-store
-  (let [store (buffer-store (memory-store))]
-    (test-block-store
-      "buffer-store" store
-      :max-size 1024
-      :blocks 25)))
+  (tests/check-store! #(buffer-store (memory-store))))