+++ test/blocks/store/memory_test.clj
@@ -2,13 +2,9 @@
   (:require
     (blocks.store
       [memory :refer [memory-store]]
-      [tests :refer [test-block-store]])
+      [tests :as tests])
     [clojure.test :refer :all]))
 
 
 (deftest ^:integration test-memory-store
-  (let [store (memory-store)]
-    (test-block-store
-      "memory-store" store
-      :max-size 1024
-      :blocks 25)))
+  (tests/check-store! memory-store))